# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: ATape HTTP API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    First-party ATape Web and CLI contract. Domain behavior is owned by the
    Authentication, Team, ingestion, conversation, Raw, and Search Modules;
    this document describes only their HTTP representation.
servers:
  - url: https://api.atape.dev
    description: Official service API origin
...
paths:
...
  /api/v1/sessions/{sessionId}:
    parameters:
      - {$ref: "#/components/parameters/SessionId"}
    get:
      operationId: getCapturedSession
      tags: [Captured sessions]
      summary: Read a Canonical conversation
      x-atape-route: {class: web_only, requestBodyPolicy: none, requestBodyLimit: 0, cachePolicy: no-store, originPolicy: none, freshAuthentication: false, requiresIdempotencyKey: false}
      security: [{webSession: []}]
      parameters:
        - {name: thread, in: query, required: false, schema: {type: string}}
      responses:
        "200": {$ref: "#/components/responses/GenericJSON"}
        "4XX": {$ref: "#/components/responses/Problem"}
        default: {$ref: "#/components/responses/Problem"}
    delete:
      operationId: deleteCapturedSession
      tags: [Captured sessions]
      summary: Delete a captured Session
      description: The capturing User may delete their own Session; deleting another User's Session requires a fresh Owner session.
      x-atape-route: {class: web_only, requestBodyPolicy: none, requestBodyLimit: 0, cachePolicy: no-store, originPolicy: exact-web, freshAuthentication: false, requiresIdempotencyKey: false}
      security: [{webSession: []}]
      parameters:
        - {$ref: "#/components/parameters/Origin"}
        - {$ref: "#/components/parameters/CSRF"}
      responses:
        "204": {$ref: "#/components/responses/NoContent"}
        "4XX": {$ref: "#/components/responses/Problem"}
        default: {$ref: "#/components/responses/Problem"}
...
components:
  securitySchemes:
    webSession:
      type: apiKey
      in: cookie
      name: __Secure-atape_session
      description: The production Cookie name is derived from topology and may use the __Host- prefix.
    cliCredential:
      type: http
      scheme: bearer
      bearerFormat: atc_v1 opaque credential
  parameters:
    Origin:
      name: Origin
      in: header
      required: true
      schema: {type: string, format: uri}
      description: Must exactly equal the configured canonical Web origin.
...
    CSRF:
      name: X-ATape-CSRF
      in: header
      required: true
      schema: {type: string, minLength: 1}
      description: Session-bound synchronizer token from the Session bootstrap response.
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: {$ref: "#/components/schemas/OpaqueId"}
...
  responses:
    Problem:
      description: RFC 9457 Problem Details. The stable code determines client behavior.
      headers:
        X-Request-ID: {$ref: "#/components/headers/RequestId"}
        Cache-Control: {$ref: "#/components/headers/NoStore"}
        Retry-After: {$ref: "#/components/headers/RetryAfter"}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    NoContent:
      description: The desired state has been reached.
      headers:
        X-Request-ID: {$ref: "#/components/headers/RequestId"}
        Cache-Control: {$ref: "#/components/headers/NoStore"}
...
    GenericJSON:
      description: Successful typed JSON representation; see the focused API document for the Module schema.
      headers:
        X-Request-ID: {$ref: "#/components/headers/RequestId"}
        Cache-Control: {$ref: "#/components/headers/NoStore"}
      content:
        application/json:
          schema: {type: object}
...
  headers:
    RequestId:
      required: true
      schema: {type: string, format: uuid}
    NoStore:
      required: true
      schema: {type: string, const: no-store}
...
    RetryAfter:
      required: false
      schema: {type: integer, minimum: 1}
      description: Present for retryable 409, 429, and 503 outcomes.
  schemas:
...
    OpaqueId:
      type: string
      minLength: 1
      maxLength: 200
      description: Opaque identifier without embedded authority.
...
    ProblemField:
      type: object
      additionalProperties: false
      required: [field, code]
      properties:
        field: {type: string}
        code: {type: string}
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, code, detail, instance, requestId]
      properties:
        type: {type: string, format: uri, pattern: "^https://atape[.]dev/problems/v1/"}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        code:
          type: string
          enum:
            - invalid_request
            - invalid_user_code
            - invalid_join_code
            - invalid_device_code
            - unsupported_protocol_version
            - ambiguous_credentials
            - unauthenticated
            - session_expired
            - session_revoked
            - fresh_authentication_required
            - origin_rejected
            - csrf_rejected
            - access_denied
            - membership_role_denied
            - credential_capability_denied
            - not_found
            - method_not_allowed
            - external_identity_conflict
            - last_owner_required
            - grant_already_decided
            - grant_consumed
            - idempotency_conflict
            - idempotency_in_progress
            - resource_state_conflict
            - login_expired
            - expired_token
            - request_too_large
            - unsupported_media_type
            - validation_failed
            - slow_down
            - too_many_code_attempts
            - internal_error
            - provider_unavailable
            - service_unavailable
        detail: {type: string}
        instance: {type: string, format: uri, pattern: "^urn:atape:request:"}
        requestId: {type: string, format: uuid}
        errors:
          type: array
          items: {$ref: "#/components/schemas/ProblemField"}